Extract LoadStrategy to a separate file and add Mapper::Relation#load
